requirements.txt
# Core dependencies
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# LangGraph / LangChain orchestration
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-anthropic>=0.1.0

# HTTP & async
httpx>=0.25.0
aiohttp>=3.9.0

# Data processing
pyarrow>=14.0.0
python-dateutil>=2.8.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Calibration
netcal>=1.3.0  # probability calibration tools

# Optional: BigQuery (uncomment if using live GDELT)
# google-cloud-bigquery>=3.13.0
# google-auth>=2.25.0

# Optional: TimesFM (uncomment if using Google's TSFM)
# timesfm>=1.0.0

# Optional: Chronos (uncomment if using Amazon's TSFM)
# chronos-forecasting>=1.0.0

# Serialization & config
pydantic>=2.5.0
python-dotenv>=1.0.0
typer>=0.9.0
rich>=13.7.0
